Move Shutdown and Idle jobs under ConsumerGroups namespace
Share groups (KIP-932) don't have the same partition ownership lifecycle
as consumer groups — no revocations, no long-lived consumer instances.
Shutdown and Idle are consumer-group-specific lifecycle jobs, so they
belong under Processing::ConsumerGroups::Jobs alongside Consume, Eofed,
and Revoked.

Only Jobs::Base remains as the generic base class under Processing::Jobs.
Backwards-compatible aliases are kept at the old paths.
