+++ spec/lib/karafka/instrumentation/logger_listener_spec.rb
@@ -94,7 +94,7 @@ RSpec.describe_current do
   describe "#on_worker_process" do
     subject(:trigger) { listener.on_worker_process(event) }
 
-    let(:job) { Karafka::Processing::Jobs::Shutdown.new(executor) }
+    let(:job) { Karafka::Processing::ConsumerGroups::Jobs::Shutdown.new(executor) }
     let(:executor) { build(:processing_executor) }
     let(:payload) { { job: job } }
 
@@ -104,7 +104,7 @@ RSpec.describe_current do
   describe "#on_worker_processed" do
     subject(:trigger) { listener.on_worker_processed(event) }
 
-    let(:job) { Karafka::Processing::Jobs::Shutdown.new(executor) }
+    let(:job) { Karafka::Processing::ConsumerGroups::Jobs::Shutdown.new(executor) }
     let(:executor) { build(:processing_executor) }
     let(:payload) { { job: job, time: 2 } }
